Player classes, rolls, stats and modifiers
Added a name, description, hit die, primary ability stat, and save for each class. Changed how player rolls are created and used. Now each roll is individually added and when it is used, it is then removed from the list of available rolls to give to stats. Modifiers work the same way.